feat: api helper 도입 및 코드 동기화
